pyproject.toml
[project]
name = "x86-emulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "capstone>=5.0.7",
    "pyelftools>=0.32",
]
